feat: 마커 DOM 유틸 libs로 이동 및 테스트 추가
- createCustomMarkerElement를 libs로 이동하고 도메인 re-export 유지
- 유틸 테스트 추가
